+++ tests/testthat/test-makeLongData.R
@@ -159,7 +159,7 @@ test_that("collapsing long data works", {
                              metricGrouping = "---", idCol = "Method",
                              metricGroupCol = "metricGroup", 
                              valueCol = "ScaledValue", weightCol = "Weight", 
-                             metricCol = "Metric")
+                             metricCol = "Metric", collapseMethod = "mean")
     expect_equal(ldw, cld)
     
     ## With grouping
@@ -182,7 +182,7 @@ test_that("collapsing long data works", {
                              metricGrouping = "num", idCol = "Method",
                              metricGroupCol = "metricGroup", 
                              valueCol = "ScaledValue", weightCol = "Weight", 
-                             metricCol = "Metric")
+                             metricCol = "Metric", collapseMethod = "mean")
     expect_s3_class(cld, "data.frame")
     expect_equal(dim(cld), c(6, 5))
     expect_named(cld, c("Method", "metricGroup", "ScaledValue", "Weight", "Metric"))
@@ -199,5 +199,123 @@ test_that("collapsing long data works", {
                  mean(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 2]))
     expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 2], 
                  mean(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 2]))
+    expect_equal(cld$Weight, c(0.7, 0.1, 0.7, 0.1, 0.7, 0.1))
+    
+    ## With grouping - max collapse
+    ld <- .makeLongData(df = df, idCol = "Method", 
+                        metrics = c("m1", "m2", "m3"), 
+                        metricCol = "Metric", valueCol = "ScaledValue",
+                        metricGrouping = "num", metricInfo = metricInfo, 
+                        metricGroupCol = "metricGroup")
+    ldw <- .addWeightsToLongData(df = ld, metricCollapseGroup = TRUE, 
+                                 metricGrouping = "num",
+                                 metricGroupCol = "metricGroup", 
+                                 weights = list(m1_weight = 0.2, m2_weight = 0.3, m3_weight = 0.1,
+                                                Metric_m1_weight = 0.1, Metric_m2_weight = 0.5,
+                                                Metric_m3_weight = 0.7, num_1_weight = 0.7, 
+                                                num_2_weight = 0.1, lets3_m_weight = 0.4, 
+                                                lets3_n_weight = 0.1, lets3_o_weight = 0.9),
+                                 weightCol = "Weight", metricCol = "Metric",
+                                 metrics = c("m1", "m2", "m3"))
+    cld <- .collapseLongData(df = ldw, metricCollapseGroup = TRUE, 
+                             metricGrouping = "num", idCol = "Method",
+                             metricGroupCol = "metricGroup", 
+                             valueCol = "ScaledValue", weightCol = "Weight", 
+                             metricCol = "Metric", collapseMethod = "max")
+    expect_s3_class(cld, "data.frame")
+    expect_equal(dim(cld), c(6, 5))
+    expect_named(cld, c("Method", "metricGroup", "ScaledValue", "Weight", "Metric"))
+    expect_equal(cld$metricGroup, cld$Metric)
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 1], 
+                 max(ldw$ScaledValue[ldw$Method == "A" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 1], 
+                 max(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 1], 
+                 max(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 2], 
+                 max(ldw$ScaledValue[ldw$Method == "A" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 2], 
+                 max(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 2], 
+                 max(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 2]))
+    expect_equal(cld$Weight, c(0.7, 0.1, 0.7, 0.1, 0.7, 0.1))
+    
+    ## With grouping - min collapse
+    ld <- .makeLongData(df = df, idCol = "Method", 
+                        metrics = c("m1", "m2", "m3"), 
+                        metricCol = "Metric", valueCol = "ScaledValue",
+                        metricGrouping = "num", metricInfo = metricInfo, 
+                        metricGroupCol = "metricGroup")
+    ldw <- .addWeightsToLongData(df = ld, metricCollapseGroup = TRUE, 
+                                 metricGrouping = "num",
+                                 metricGroupCol = "metricGroup", 
+                                 weights = list(m1_weight = 0.2, m2_weight = 0.3, m3_weight = 0.1,
+                                                Metric_m1_weight = 0.1, Metric_m2_weight = 0.5,
+                                                Metric_m3_weight = 0.7, num_1_weight = 0.7, 
+                                                num_2_weight = 0.1, lets3_m_weight = 0.4, 
+                                                lets3_n_weight = 0.1, lets3_o_weight = 0.9),
+                                 weightCol = "Weight", metricCol = "Metric",
+                                 metrics = c("m1", "m2", "m3"))
+    cld <- .collapseLongData(df = ldw, metricCollapseGroup = TRUE, 
+                             metricGrouping = "num", idCol = "Method",
+                             metricGroupCol = "metricGroup", 
+                             valueCol = "ScaledValue", weightCol = "Weight", 
+                             metricCol = "Metric", collapseMethod = "min")
+    expect_s3_class(cld, "data.frame")
+    expect_equal(dim(cld), c(6, 5))
+    expect_named(cld, c("Method", "metricGroup", "ScaledValue", "Weight", "Metric"))
+    expect_equal(cld$metricGroup, cld$Metric)
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 1], 
+                 min(ldw$ScaledValue[ldw$Method == "A" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 1], 
+                 min(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 1], 
+                 min(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "A" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 2]))
+    expect_equal(cld$Weight, c(0.7, 0.1, 0.7, 0.1, 0.7, 0.1))
+    
+    ## With grouping - min collapse (with NAs)
+    df0 <- df
+    df0$m1[1] <- df0$m2[1] <- NA
+    ld <- .makeLongData(df = df0, idCol = "Method", 
+                        metrics = c("m1", "m2", "m3"), 
+                        metricCol = "Metric", valueCol = "ScaledValue",
+                        metricGrouping = "num", metricInfo = metricInfo, 
+                        metricGroupCol = "metricGroup")
+    ldw <- .addWeightsToLongData(df = ld, metricCollapseGroup = TRUE, 
+                                 metricGrouping = "num",
+                                 metricGroupCol = "metricGroup", 
+                                 weights = list(m1_weight = 0.2, m2_weight = 0.3, m3_weight = 0.1,
+                                                Metric_m1_weight = 0.1, Metric_m2_weight = 0.5,
+                                                Metric_m3_weight = 0.7, num_1_weight = 0.7, 
+                                                num_2_weight = 0.1, lets3_m_weight = 0.4, 
+                                                lets3_n_weight = 0.1, lets3_o_weight = 0.9),
+                                 weightCol = "Weight", metricCol = "Metric",
+                                 metrics = c("m1", "m2", "m3"))
+    cld <- .collapseLongData(df = ldw, metricCollapseGroup = TRUE, 
+                             metricGrouping = "num", idCol = "Method",
+                             metricGroupCol = "metricGroup", 
+                             valueCol = "ScaledValue", weightCol = "Weight", 
+                             metricCol = "Metric", collapseMethod = "min")
+    expect_s3_class(cld, "data.frame")
+    expect_equal(dim(cld), c(6, 5))
+    expect_named(cld, c("Method", "metricGroup", "ScaledValue", "Weight", "Metric"))
+    expect_equal(cld$metricGroup, cld$Metric)
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 1], NA_real_)
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 1], 
+                 min(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 1], 
+                 min(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 1]))
+    expect_equal(cld$ScaledValue[cld$Method == "A" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "A" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "B" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "B" & ldw$metricGroup == 2]))
+    expect_equal(cld$ScaledValue[cld$Method == "C" & cld$metricGroup == 2], 
+                 min(ldw$ScaledValue[ldw$Method == "C" & ldw$metricGroup == 2]))
     expect_equal(cld$Weight, c(0.7, 0.1, 0.7, 0.1, 0.7, 0.1))
 })